split activity log modal into its own component and moved a few functions with it
